fix: redirect to homepage if not yet logged in
Bug:
When a visitor goes to `/app`, and are not yet logged in, they are not redirected to the homepage.

Reason:
`shouldRedirect`, a value set to prevent redirects when the `useAuth` hook is doing its 'magic', is not set to true after the 'magic' is done, if the user is not logged in.

Fix:
Set `shouldRedirect` to true after the 'magic' is done.