fix: improve logout performance on slow networks
Bug:
Logging out of the application is dependent on a network request which clears httponly cookies.

Workaround:
Logging out now changes to the logged out ui optimistically.

Issues with workaround:
When logging out, if the network request fails then the user's data will not be cleared. This is especially problematic on public computers or if a person is logging in with a different account. These are edge cases and for now I am choosing not to address them.